image_processing/find_regions/requirements.txt
pika
Pillow
jsonpath-rw
dpath
numpy
opencv-python
